backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
pillow==10.2.0
python-dotenv==1.0.0
